Add tests for CADET paths and CADET Runner attributes
